Create b.cpp
In this code, we define a Course class with three member variables: name, code, and duration. We also define a parameterized constructor that takes these three variables as inputs and initializes them.

We then define a member function getCourseDetails that takes a course code as input and checks if the course code matches the code of the current course object. If there is a match, the function prints the course details using cout statements. Otherwise, it prints a message saying that the course with the given code was not found.

In the main function, we create a Course object and add its details using the parameterized constructor. We then call the getCourseDetails member function on the course object to get its details by course code. Finally, we print the course details using cout statements.