Create a.cpp
In this code, we define a UGCourse class as the base class, and PCourse and DiplomaCourse classes as derived classes that inherit from the UGCourse class. We use the public access specifier to make the inherited members of the UGCourse class accessible to the derived classes.

We define a constructor for each derived class that initializes its own member variables as well as the inherited member variables using the base class constructor.

We also define a printCourse function for each class that prints the course details using cout statements. In the PCourse and DiplomaCourse classes, we override the printCourse function of the base class to include the additional member variables specific to each class.

In the main function, we create objects of each class and print their details using the printCourse function.